V3.1.1 enhancements : -Use DB with SQLITE as main storage of the application. this will reduce the dependancy of files, and increase the number of features that we have with SQLITE that we can add it to the application later easily, like Sorting/Filer/Grouping/relation with other tables like categories and rating, add more station metadata, and many more. 	-adding more metadata to stations (currently no way to change all metadata throw UI, we can only import xml with all metadata of channel, we may add editing later in next versions). 	-new station rating modules (currently it's local rating only, not sync to central rating, may be later in next versions) 	-Using Facebook fresco image viewer (SimpleDraweeView) with it's great loading features 	-Apply Material design components and transitions 	-Integrate with Firebase compoenents 			-Authentication 			-Analytics 	-Add login by Google account function (currently nothing happend after login, but in the next versions we can connect user by ID and save his data to cloud for example, and many more.)

+++ app/src/main/java/org/y20k/transistor/helpers/SingletonProperties.java
@@ -12,7 +12,7 @@ public class SingletonProperties {
 
     public PlaybackStatus CurrentSelectedStation_Playback_Status = PlaybackStatus.STOPPED;
     public String CurrentStation_ID = null;
-    public int CurrentSelectedStation_ID = -1;
+    public long CurrentSelectedStation_ID = -1;
 
     private static Context mContext;
 
@@ -39,10 +39,10 @@ public class SingletonProperties {
         return mInstance;
     }
 
-    public int getLastRunningStation_ID() {
+    public long getLastRunningStation_ID() {
         //get from application setting/ preferences
         SharedPreferences settings = PreferenceManager.getDefaultSharedPreferences(mContext);
-        return settings.getInt(TransistorKeys.PREF_STATION_ID_LAST, -1);
+        return settings.getLong(TransistorKeys.PREF_STATION_ID_LAST, -1);
     }
 
     /**
@@ -52,11 +52,11 @@ public class SingletonProperties {
     public boolean getIsPlayback() {
         return (CurrentSelectedStation_Playback_Status==PlaybackStatus.LOADING || CurrentSelectedStation_Playback_Status==PlaybackStatus.PLAYING);
     }
-    public void setLastRunningStation_ID(int lastRunningStation_ID) {
+    public void setLastRunningStation_ID(long lastRunningStation_ID) {
         //add to application setting/ preferences
         SharedPreferences settings = PreferenceManager.getDefaultSharedPreferences(mContext);
         SharedPreferences.Editor editor = settings.edit();
-        editor.putInt(TransistorKeys.PREF_STATION_ID_LAST, lastRunningStation_ID);
+        editor.putLong(TransistorKeys.PREF_STATION_ID_LAST, lastRunningStation_ID);
         editor.apply();
     }
 }
